Discard Apollo cache on each page transition
