+++ frontend/components/CompletionListItem.tsx
@@ -12,18 +12,28 @@ import DoneIcon from "@material-ui/icons/Done"
 import Avatar from "@material-ui/core/Avatar"
 import { CardTitle, CardSubtitle } from "components/Text/headers"
 import { addDomain } from "/util/imageUtils"
+import Link from "next/link"
 
 const StyledButton = styled(Button)`
   height: 50%;
-  margin: auto;
+
   color: black;
+`
+
+const StyledA = styled.a`
+  margin: auto;
 `
 
 const RegisterCompletionButton = ({ course }: { course: string }) => {
   return (
-    <StyledButton color="secondary" href={`/register-completion/${course}`}>
-      Register Completion
-    </StyledButton>
+    <Link
+      href="/register-completion/[slug]"
+      as={`/register-completion/${course}`}
+    >
+      <StyledA>
+        <StyledButton color="secondary">Register Completion</StyledButton>
+      </StyledA>
+    </Link>
   )
 }
