+++ frontend/lib/with-apollo-client/fetch-user-details.ts
@@ -1,0 +1,22 @@
+import gql from "graphql-tag"
+import { ApolloClient, NormalizedCacheObject } from "apollo-boost"
+
+export const UserDetailQuery = gql`
+  query UserOverView {
+    currentUser {
+      id
+      first_name
+      last_name
+      email
+    }
+  }
+`
+export default async function fetchUserDetails(
+  apollo: ApolloClient<NormalizedCacheObject>,
+) {
+  const { data } = await apollo.query({
+    query: UserDetailQuery,
+    fetchPolicy: "cache-first",
+  })
+  return data.currentUser
+}
